Rename all-editor-modes.js to _all.js
